Refatora gerenciamento de aop

+++ src/scielo/bin/xml/modules/dbm_isis.py
@@ -409,20 +409,21 @@ class IsisDAO(object):
             base = temp_file.name
             self.cisis.search(db_filename, expr, base)
 
+        r = []
         id_filename = base + '.id'
-
-        self.cisis.i2id(base, id_filename)
-        r = IDFile().read(id_filename)
-        if len(r) > 0:
-            if temp_file is not None:
+        if os.path.isfile(base + '.mst'):
+            self.cisis.i2id(base, id_filename)
+            r = IDFile().read(id_filename)
+            if len(r) > 0:
+                if temp_file is not None:
+                    try:
+                        os.unlink(temp_file.name)
+                    except:
+                        pass
                 try:
-                    os.unlink(temp_file.name)
+                    os.unlink(id_filename)
                 except:
                     pass
-            try:
-                os.unlink(id_filename)
-            except:
-                pass
         return r
 
     def get_id_records(self, id_filename):
